requirements.txt
tensorflow
pandas
numpy
notebook
matplotlib
seaborn
python-box
pyYAML
pqdm
ensure
joblib
type-pyYAML
scipy
-e .